+++ day2/2.2/UITabController/UITabController/AppDelegate.swift
@@ -15,14 +15,31 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
+        
+        var viewControllers: [UIViewController] = []
+        
+        let firstViewController = FirstViewController()
+        firstViewController.tabBarItem = UITabBarItem(tabBarSystemItem: UITabBarSystemItem.mostRecent, tag: 1)
+        viewControllers.append(firstViewController)
+        
+        let secondViewController = SecondViewController()
+        secondViewController.tabBarItem = UITabBarItem(tabBarSystemItem: UITabBarSystemItem.mostViewed, tag: 2)
+        viewControllers.append(secondViewController)
+        
+        let thirdViewController = ThirdViewController()
+        thirdViewController.tabBarItem = UITabBarItem(tabBarSystemItem: UITabBarSystemItem.contacts, tag: 3)
+        viewControllers.append(thirdViewController)
+        
+        let tabBarController = UITabBarController()
+        tabBarController.setViewControllers(viewControllers, animated: false)
+
         window = UIWindow(frame: UIScreen.main.bounds)
         window?.makeKeyAndVisible()
-//        let mainTabController: MainTabController = MainTabController
-        window?.rootViewController = MainTabController
+        window?.rootViewController = tabBarController
         
         return true
     }
-
+}
     func applicationWillResignActive(_ application: UIApplication) {
         // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
         // Use this method to pause ongoing tasks, disable timers, and invalidate graphics rendering callbacks. Games should use this method to pause the game.
@@ -46,5 +63,5 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
 
-}
+
