+++ day2/2.2/UITabController/UITabController/FirstViewController.swift
@@ -1,26 +1,13 @@
-//
-//  ViewController.swift
-//  UITabController
-//
-//  Created by Kunio Terada on 2018/08/29.
-//  Copyright © 2018年 Kunio Terada. All rights reserved.
-//
-
 import UIKit
 
 class FirstViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-        self.view.backgroundColor = UIColor.red
+        view.backgroundColor = UIColor.white
     }
 
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
     }
-
-
 }
-
