+++ day2/2.2/UITabController/UITabController/MainTabController.swift
@@ -1,38 +1,14 @@
+//import UIKit
 //
-//  MainTabController.swift
-//  UITabController
+//class MainTabController: UIViewController {
+//    
+//    override func viewDidLoad() {
+//        super.viewDidLoad()
+//        
+//    
+//    override func didReceiveMemoryWarning() {
+//        super.didReceiveMemoryWarning()
+//    }
 //
-//  Created by Kunio Terada on 2018/08/29.
-//  Copyright © 2018年 Kunio Terada. All rights reserved.
+//}
 //
-
-import UIKit
-
-class MainTabController: UITabBarController {
-    
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        
-        var viewControllers: [UIViewController] = []
-        
-        let firstViewController = FirstViewController()
-        firstViewController.tabBarItem = UITabBarItem(tabBarSystemItem: UITabBarSystemItem.mostRecent, tag: 1)
-        viewControllers.append(firstViewController)
-        
-        let secondViewController = SecondViewController()
-        secondViewController.tabBarItem = UITabBarItem(tabBarSystemItem: UITabBarSystemItem.mostViewed, tag: 2)
-        viewControllers.append(secondViewController)
-        
-        let thirdViewController = ThirdViewController()
-        thirdViewController.tabBarItem = UITabBarItem(tabBarSystemItem: UITabBarSystemItem.contacts, tag: 3)
-        viewControllers.append(thirdViewController)
-        
-        self.setViewControllers(viewControllers, animated: false)
-        
-        
-        // なぜか0だけだと選択されないので1にしてから0に
-        self.selectedIndex = 1
-        self.selectedIndex = 0
-    }
-}
-
