+++ day2/2.2/UITabController/UITabController/SecondViewController.swift
@@ -1,26 +1,13 @@
-//
-//  NextViewController.swift
-//  UITabController
-//
-//  Created by Kunio Terada on 2018/08/29.
-//  Copyright © 2018年 Kunio Terada. All rights reserved.
-//
-
 import UIKit
 
 class SecondViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-        self.view.backgroundColor = UIColor.blue
-        
+        view.backgroundColor = UIColor.gray
     }
     
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
     }
-    
-    
 }
